Add explicit cast to disambiguate pushKV calls.
Where pushKV is used with size_t, explicitly cast to int.  This resolves
compiler errors on win32 and mac platforms.

+++ src/rpc/misc.cpp
@@ -532,7 +532,7 @@ getstatsforheight (const JSONRPCRequest& request)
   assert (nNameTx + nCurrencyTx == block.vtx.size ());
   transactions.pushKV ("currency", static_cast<int> (nCurrencyTx));
   transactions.pushKV ("name", static_cast<int> (nNameTx));
-  transactions.pushKV ("game", vGameTx.size ());
+  transactions.pushKV ("game", static_cast<int> (vGameTx.size ()));
 
   GameState gameState(Params ().GetConsensus ());
   if (!pgameDb->get (block.GetHash (), gameState))
